Move alllll the sync code to sync.js - still more to do (in FIXME XXX)
